feat: open source issues and add it to runs tab
