+++ packages/server/api/src/app/app.ts
@@ -38,7 +38,7 @@ import { platformRunHooks } from './ee/flow-run/cloud-flow-run-hooks'
 import { platformFlowTemplateModule } from './ee/flow-template/platform-flow-template.module'
 import { globalConnectionModule } from './ee/global-connections/global-connection-module'
 import { emailService } from './ee/helper/email/email-service'
-import { issuesModule } from './ee/issues/issues-module'
+import { issuesModule } from './flows/issues/issues-module'
 import { licenseKeysModule } from './ee/license-keys/license-keys-module'
 import { managedAuthnModule } from './ee/managed-authn/managed-authn-module'
 import { mcpModule } from './ee/mcp/mcp-module'
