+++ packages/server/api/src/app/database/database-connection.ts
@@ -20,7 +20,7 @@ import { AppSumoEntity } from '../ee/billing/appsumo/appsumo.entity'
 import { ConnectionKeyEntity } from '../ee/connection-keys/connection-key.entity'
 import { CustomDomainEntity } from '../ee/custom-domains/custom-domain.entity'
 import { FlowTemplateEntity } from '../ee/flow-template/flow-template.entity'
-import { IssueEntity } from '../ee/issues/issues-entity'
+import { IssueEntity } from '../flows/issues/issues-entity'
 import { MCPEntity } from '../ee/mcp/mcp-entity'
 import { OAuthAppEntity } from '../ee/oauth-apps/oauth-app.entity'
 import { PlatformBillingEntity } from '../ee/platform-billing/platform-billing.entity'
