+++ packages/server/api/src/app/ee/alerts/alerts-handler.ts
@@ -25,10 +25,6 @@ async function scheduleSendingReminder(params: IssueRemindersParams, log: Fastif
     if (params.issueCount === 1) {
         const project = await projectService.getOneOrThrow(projectId)
         const platform = await platformService.getOneOrThrow(project.platformId)
-        if (!platform.flowIssuesEnabled) {
-            return
-        }
-        
         const reminderKey = `reminder:${projectId}`
         const isEmailScheduled = await getRedisConnection().get(reminderKey)
         if (isEmailScheduled) {
