+++ packages/server/api/src/app/flows/issues/issues-service.ts
@@ -4,7 +4,7 @@ import { ActivepiecesError, ApId, apId, ErrorCode, isNil, SeekPage, spreadIfDefi
 import dayjs from 'dayjs'
 import { FastifyBaseLogger } from 'fastify'
 import { repoFactory } from '../../core/db/repo-factory'
-import { flowVersionService } from '../../flows/flow-version/flow-version.service'
+import { flowVersionService } from '../flow-version/flow-version.service'
 import { buildPaginator } from '../../helper/pagination/build-paginator'
 import { paginationHelper } from '../../helper/pagination/pagination-utils'
 import { telemetry } from '../../helper/telemetry.utils'
